fix: api extra hooks test ci failure
